Allow weekly build to merge before push
Fixes https://github.com/jenkins-infra/release/issues/387

Intentionally limited to only push if using the weekly release profile
because we do not expect to receive commits during a build on a branch
that is not a weekly build.  LTS builds and security builds that have
a commit arrive during the build should continue to fail the build.