Use Maven release plugin 3.1.1 (#782)
* Revert "Clear the changelist in a spotless compatible format (#781)"

Fails the build because `mvn release:prepare` requires that the changelist
must be a SNAPSHOT version.

This reverts commit fbf53386a937156ced18dca23057d8d8da600d04.

* Use Maven release plugin 3.1.1

Maven release plugin 3.2.0 (included in parent pom 1.142) seems to
ignore our spotless.check.skip definition.  It is checking the spotless
formatting and fails that check when `mvn release:prepare` creates
the empty changelist tag in a format that is not the format accepted
by spotless.